Rename: 공통 서비스 이름 변경
